first swing at abstract class
